Fix blendMode sign-shift in CalcChara (71.63->71.76%)
Same logical-shift fix as PCAnimCtrl: the SetAnim blend-mode argument
-1 - (m_currentAnimIndex >> 31) uses a logical (srwi) shift in the
original. Behavior-identical.

+++ src/wm_menu.cpp
@@ -7466,7 +7466,8 @@ void CMenuPcs::CalcChara()
 				animState[0] = 0;
 				animState[1] = -1;
 				animState[2] = rand() % 250;
-				loadHandle->SetAnim(baseAnim, -1, -1, -1 - (loadHandle->m_currentAnimIndex >> 31), 1);
+				loadHandle->SetAnim(baseAnim, -1, -1,
+				    -1 - static_cast<int>(static_cast<unsigned int>(loadHandle->m_currentAnimIndex) >> 31), 1);
 				animState[3] = reinterpret_cast<int*>(reinterpret_cast<unsigned char*>(loadHandle->m_model) + 0xB4)[0];
 				animState[4] = reinterpret_cast<int*>(reinterpret_cast<unsigned char*>(loadHandle->m_model) + 0xC0)[0];
 			}
